+++ bridge/src/main/scala/bridge/JSArray.scala
@@ -1,7 +1,8 @@
 package bridge
 
+import korolev.async.Async
+
 import scala.language.higherKinds
-import korolev.Async
 
 /**
  * JavaScript Array presentation
