+++ bridge/src/main/scala/bridge/JSLink.scala
@@ -1,7 +1,8 @@
 package bridge
 
+import korolev.async.Async
+
 import scala.language.higherKinds
-import korolev.Async
 
 /**
  * Link to entity on page side. By default, all links
