+++ server/src/main/scala/korolev/server/JsonQueuedJsAccess.scala
@@ -3,8 +3,8 @@ package korolev.server
 import java.util.concurrent.ConcurrentLinkedQueue
 
 import bridge.JSAccess
-import korolev.Async
 import korolev.Async.Promise
+import korolev.async.Async
 
 import scala.annotation.switch
 import scala.collection.concurrent.TrieMap
